+++ test/dummy_app/db/schema.rb
@@ -10,6 +10,6 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20120221152352) do
+ActiveRecord::Schema.define(version: 20120221152352) do
 
 end