+++ test/integration/prawn_rails_test.rb
@@ -14,50 +14,51 @@ class NavigationTest < ActionDispatch::IntegrationTest
     assert ActionView::Template::Handlers.extensions.include?(:prawn)
   end
 
-  test "render html action" do
+  test "Renders html action" do
     get '/reports/sample'
     assert_response :success
-    assert @response.body.include?("<h1>Hello World!</h1>")
+    #assert @response.body.include?("<h1>Hello World!</h1>")
+    assert_match("<h1>Hello World!</h1>", @response.body)
   end
 
-  test "render pdf action" do
+  test "Renders pdf to string" do
+    pdf_str = ApplicationController.new.render_to_string("reports/sample.pdf", locals: {:@items => []})
+
+    reader = PDF::Reader.new(StringIO.new(pdf_str))
+    assert_not_nil(reader)
+
+    page_str = reader.pages[0].to_s
+    assert page_str.include?('Hello World!')
+    assert_not page_str.include?("<h1>Hello World!</h1>")
+  end
+
+  test "Renders sample pdf action" do
     get '/reports/sample', params: {format: :pdf}
     assert_response :success
     assert_not @response.body.include?("<h1>Hello World!</h1>")
+
+    reader = PDF::Reader.new(StringIO.new(@response.body))
+    assert_not_nil(reader)
+
+    assert_equal(1, reader.page_count)
+
+    page_str = reader.pages[0].to_s
+
+    assert page_str.include?('Hello World!')
+    assert_not page_str.include?("<h1>Hello World!</h1>")
   end
 
-  test "Missing pdf format" do
-    assert true
-  end
-
-  test "strip_tags" do
-    assert true
-  end
-
-  test "ActionView::Base.full_sanitizer.sanitize(html_string)" do
-    assert true
-  end
-
-  test "" do
-    assert true
-  end
-
-  test "legacy test - render pdf to string" do
-    ApplicationController.new.render_to_string("reports/sample.pdf", locals: {'@items' => []})
-  end
-
-  test "legacy test - render pdf" do
-    get '/reports/sample', params: {format: :pdf}
+  test "Renders table pdf action using auto-required plugin Prawn-Table" do
+    get '/reports/table', params: {format: :pdf}
     assert_response :success
 
     reader = PDF::Reader.new(StringIO.new(@response.body))
     assert_not_nil(reader)
 
     assert_equal(1, reader.page_count)
-    assert_match(/Hello There :D/, reader.pages[0].to_s)
-    assert_no_match(/This should not be any where/, reader.pages[0].to_s)
-    assert_no_match(/ITEMS ARE nil/, reader.pages[0].to_s)
-    assert_match(/with html<bold><\/bold>/,reader.pages[0].to_s)
-    assert_match(/without html LD/,reader.pages[0].to_s)
+
+    page_str = reader.pages[0].to_s
+
+    assert_equal(page_str, "1 2 3\n\n4 5 6\n\n7 8 9")
   end
 end