+++ test/prawn_rails_test.rb
@@ -4,14 +4,8 @@ require 'test_helper'
 
 class PrawnRailsTest < ActiveSupport::TestCase
 
-  # TODO remove this test
-  include PrawnRails::Extension
-  test "html_strip" do
-    assert_nil html_strip(nil)
-    assert_not_nil html_strip("")
-  end
-
   test "config" do
     assert true
   end
+
 end