+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 * `v1.1.0` - UNRELEASED
-  - Breaking Change - Remove `html_strip` helper method. Use `ActionView::Base.full_sanitizer.sanitize(html_string)` if you were using this feature.
   - [#29](https://github.com/cortiz/prawn-rails/pull/29) - Require all installed prawn extensions
+  - Removed unnecessary `html_strip` helper method. If you were using this feature, instead use Rails built in `strip_tags(html_str)` or `ActionView::Base.full_sanitizer.sanitize(html_str)`
   - Add legitimate tests
 
 * `v1.0.1`